refactor(usage): drop the View suffix, and let the page own the fetching
UsageBreakdown is now the component: five props in, no queries. The wiring
moved up to UsageDashboardPage, which is allowed to fetch, so nothing needs a
suffix to distinguish it from the part that renders.

The dashboard takes the breakdown through a slot, the same way it already
takes filters and for the same reason: what goes in there fetches per key. It
also sheds organisationId, billingPeriod and projectId, which it only ever
forwarded.

Both story sets still work. UsageBreakdown renders from props, and the
dashboard stories omit the slot.
