+++ frontend/web/components/pages/usage/UsageDashboard.tsx
@@ -1,9 +1,7 @@
 import { FC, ReactNode } from 'react'
-import { Req } from 'common/types/requests'
 import { Res } from 'common/types/responses'
 import { PlanLimit } from 'components/shared/UsageBar/utils'
 import EmptyState from 'components/EmptyState'
-import UsageBreakdown from './components/UsageBreakdown'
 import UsageMeter from './components/UsageMeter'
 import UsageOverTime from './components/UsageOverTime'
 
@@ -14,22 +12,18 @@ export type UsageDashboardProps = {
   hasBillingPeriod: boolean
   isError?: boolean
   isLoading?: boolean
-  organisationId: number
-  billingPeriod: Req['getOrganisationUsage']['billing_period']
-  projectId: number | undefined
   filters?: ReactNode
+  breakdown?: ReactNode
 }
 
 const UsageDashboard: FC<UsageDashboardProps> = ({
-  billingPeriod,
+  breakdown,
   data,
   filters,
   hasBillingPeriod,
   isError,
   isLoading,
   limit,
-  organisationId,
-  projectId,
   total,
 }) => {
   let content
@@ -59,12 +53,7 @@ const UsageDashboard: FC<UsageDashboardProps> = ({
           isBillingPeriod={hasBillingPeriod}
         />
 
-        <UsageBreakdown
-          organisationId={organisationId}
-          billingPeriod={billingPeriod}
-          data={data}
-          projectId={projectId}
-        />
+        {breakdown}
       </>
     )
   }
